move find_entities function to entity service
