+++ src/dispatch/signal/flows.py
@@ -3,6 +3,7 @@ from dispatch.database.core import SessionLocal
 from dispatch.project.models import Project
 from dispatch.case import service as case_service
 from dispatch.case import flows as case_flows
+from dispatch.entity import service as entity_service
 from dispatch.signal import service as signal_service
 from dispatch.tag import service as tag_service
 from dispatch.signal.models import SignalInstanceCreate, RawSignal
@@ -34,7 +35,7 @@ def create_signal_instance(
     signal_instance.signal = signal
     db_session.commit()
 
-    entities = signal_service.find_entities(
+    entities = entity_service.find_entities(
         db_session=db_session,
         signal_instance=signal_instance,
         entity_types=signal.entity_types,
